+++ src/main/java/com/adphilip/conf/conference/repositories/SpeakerRepository.java
@@ -4,6 +4,6 @@ import com.adphilip.conf.conference.models.Speaker;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpeakerRepository extends JpaRepository<Speaker, Integer>{
+public interface SpeakerRepository extends JpaRepository<Speaker, Long>{
     
 }
